Split GameInfo class into one class per calculator
The GameInfo class was being used for all calculators.  The glicko and
elo calculators were only using the beta value, which was defaulted to
4.333 when a more sane default would be 200.  The GameInfo class has
now been split into TrueSkillGameInfo, EloGameInfo and GlickoGameInfo.
The game_info parameter of all methods has been moved to the end and
given a default of None, in which case the correct GameInfo object will
be created with sane defaults.  This means the game_info parameter is
now optional unless you want to customize your skill scale.

The EloCalculator was ignoring the k_factor of the players and has been
fixed as well.